Remove condition for enabling iptables service
The variable name and the intention behind the corresponding commit
message [1] do not match what's actually done (the iptables installation
happens in installation.yml). It might have been a mistake.

More fundamentally, however, the service should be enabled
unconditionally, as that is the main goal of this role: setting up
iptables rules, and making sure they stick around. Currently, this fails
on any system where iptables came preinstalled (or in this particular
case rather: on any system where {{ iptables_directory }} was already
set up as desired).

[1] https://github.com/adfinis-sygroup/ansible-roles/commit/97be888f578975e0ebef2100a79397e869d76570